+++ src/renderer/core/canvas/litegraph/SlotCalculations.ts
@@ -13,8 +13,6 @@ import type {
 } from '@/lib/litegraph/src/interfaces'
 import { LiteGraph } from '@/lib/litegraph/src/litegraph'
 import { isWidgetInputSlot } from '@/lib/litegraph/src/node/slotUtils'
-import { layoutStore } from '@/renderer/core/layout/store/LayoutStore'
-import type { SlotLayout } from '@/renderer/core/layout/types'
 
 export interface SlotPositionContext {
   /** Node's X position in graph coordinates */
@@ -230,61 +228,4 @@ function calculateVueSlotPosition(
   const slotCenterX = isInput ? 10 : nodeWidth - 10
 
   return [nodeX + slotCenterX, nodeY + slotCenterY]
-}
-
-/**
- * Register slot layout with the layout store for hit testing
- * @param nodeId The node ID
- * @param slotIndex The slot index
- * @param isInput Whether this is an input slot
- * @param position The slot position in graph coordinates
- */
-export function registerSlotLayout(
-  nodeId: string,
-  slotIndex: number,
-  isInput: boolean,
-  position: Point
-): void {
-  const slotKey = `${nodeId}-${isInput ? 'in' : 'out'}-${slotIndex}`
-
-  // Calculate bounds for the slot (20x20 area around center)
-  const slotSize = 20
-  const halfSize = slotSize / 2
-
-  const slotLayout: SlotLayout = {
-    nodeId,
-    index: slotIndex,
-    type: isInput ? 'input' : 'output',
-    position: { x: position[0], y: position[1] },
-    bounds: {
-      x: position[0] - halfSize,
-      y: position[1] - halfSize,
-      width: slotSize,
-      height: slotSize
-    }
-  }
-
-  layoutStore.updateSlotLayout(slotKey, slotLayout)
-}
-
-/**
- * Register all slots for a node
- * @param nodeId The node ID
- * @param context The slot position context
- */
-export function registerNodeSlots(
-  nodeId: string,
-  context: SlotPositionContext
-): void {
-  // Register input slots
-  context.inputs.forEach((_, index) => {
-    const position = calculateInputSlotPos(context, index)
-    registerSlotLayout(nodeId, index, true, position)
-  })
-
-  // Register output slots
-  context.outputs.forEach((_, index) => {
-    const position = calculateOutputSlotPos(context, index)
-    registerSlotLayout(nodeId, index, false, position)
-  })
 }
